==> test-run-1/scripts/6.parse.blast.results.Rmd <==
---
title: "process.blast"
author: "M Fisher via Eily"
date: "11/19/2021"
output: 
  html_document:
    toc: yes
---

# Description 

This script takes BLAST results and makes/adds to a custom blast reference database for ASVs that are not classified using the tree-based approach in insect. 

<br>

# Setup

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```
<br>

This script requires the following packages:
```{r message=FALSE, warning=FALSE}
library(here)
library(tidyverse)
library(magrittr)
library(cowplot)
library(janitor)
library(insect)
library(taxonomizr)
```
<br>

This script calls the following custom function:
```{r}
source(here('R','custom_lca.R'))
```
<br>

User directories
```{r}
indir  <- 'test-lane-1/data/blast/raw'
outdir <- 'test-lane-1/data/blast'
```
<br>

Pull taxonomy from insect package
```{r warning=FALSE}
worlds.taxonomy <- insect::taxonomy() 
# worlds.taxonomy <- readRDS(here('data','databases','insect_worlds_taxonomy.rds'))
```
<br>
<br>

# BF3

Files containing Blast output.
```{r}
blast_filenames     <- c('mf.lane1.bf3.merged.hashes.to.blast.txt',
                         'mf.lane1.bf3.forward.hashes.to.blast.txt',
                         'mf.lane1.bf3.reverse.hashes.to.blast.txt')
filename_identifier <- c('merged','forward','reverse')
```
<br>

Output files
```{r}
consensus_out <- 'lane1_bf3_blast_multipleIDs_taxonomy.csv'
unique_out    <- 'lane1_bf3_blast_uniqueIDs_taxonomy.csv'
all_out       <- 'lane1_bf3_blast_taxonomy.csv'
```
<br>


## Read in data

The blast results should have the headers specified, following script 4. 

The column `staxid` contains the taxa IDs assigned to the DNA sequences (identified using their hashes); a hash can be assigned a single taxon ID (yay!) or multiple taxa IDs. When we coerce the column `staxid` into a numeric format, the hashes assigned to multiple taxa IDs will go to NA. We can then filter the data set to separate the hashes assigned to multiple taxa ids and explore. 

```{r data}
for(i in seq(1,length(blast_filenames))){
  tmp_file <- read_delim(here(indir,blast_filenames[i]),col_names = c("qseqid", "sseqid", "sacc", "pident", "length", "mismatch", "gapopen", "qcovus", "qstart", "qend", "sstart", "send", "evalue", "bitscore", "staxid", "qlen", "sscinames", "sseq"), delim = "\t") %>% mutate(Ntaxid=as.numeric(staxid))
  # split data by whether or not there are multiple IDs in the `staxid` column
  tmp_NA <- tmp_file %>%
    filter(is.na(Ntaxid)) %>%
    mutate(dataset = filename_identifier[i])
  tmp_dat <- tmp_file  %>%
    filter(!is.na(Ntaxid)) %>%
    mutate(dataset = filename_identifier[i])
  # split data again, by whether or not there are multiple rows / `sseqid` per `qseqid`
  tmp_multi <- tmp_dat %>%
    group_by(qseqid) %>%
    dplyr::filter(n() != 1) %>%
    ungroup()
  tmp_dat %<>% group_by(qseqid) %>%
    dplyr::filter(n() == 1) %>%
    ungroup()
  
  if(i==1){
    blast_dat <- tmp_dat
    blast_to_process  <- tmp_multi
    blast_to_process2 <- tmp_NA
  } else{
    blast_dat %<>% bind_rows(tmp_dat)
    blast_to_process  %<>% bind_rows(tmp_multi)
    blast_to_process2 %<>% bind_rows(tmp_NA)
  }
  
}
rm(tmp_file, tmp_dat, tmp_multi, tmp_NA)

unique(blast_dat$dataset)
dim(blast_dat)
dim(blast_to_process)
dim(blast_to_process2)
```
<br>

Pull taxonomy from insect package
```{r warning=FALSE}
worlds.taxonomy <- insect::taxonomy() 
# worlds.taxonomy <- readRDS(here('data','databases','insect_worlds_taxonomy.rds'))
```
<br>
<br>

## Process results

### re-format multiple staxids & conduct LCA

extract all of the taxa IDs 
```{r}
multi_taxids <- blast_to_process2 %>%
  distinct(staxid) %>%
  pull()
multi_taxids <- unlist(sapply(multi_taxids,str_split,";")); names(multi_taxids) <- NULL
multi_taxids
```
<br>

Use insect's database to get the associated taxonomy. Coerce the output (a list) into a dataframe.
```{r}
multi_taxids_taxonomy <- insect::get_lineage(as.numeric(multi_taxids), worlds.taxonomy)

multi_taxids_taxonomy_df <- tibble (staxid = multi_taxids,
                         taxonomy = multi_taxids_taxonomy) %>%
  mutate (Kingdom = map_chr(taxonomy, ~.x["kingdom"]),
          Phylum = map_chr(taxonomy, ~.x["phylum"]),
          Class = map_chr(taxonomy, ~.x["class"]),
          Order = map_chr(taxonomy, ~.x["order"]),
          Family= map_chr(taxonomy, ~.x["family"]),
          Genus = map_chr(taxonomy, ~.x["genus"]),
          Species = map_chr(taxonomy, ~.x["species"]))
```
<br>

Attach the taxonomy information to the hashes that were assigned multiple taxIDs. The resulting data frame has one row per unique taxID per hash.
```{r}
tmp_ids <- c("id1","id2","id3","id4","id5","id6","id7","id8","id9","id10")
multi_taxids_df <- blast_to_process2 %>%
  separate(staxid, all_of(tmp_ids), fill = "right") %>%
  pivot_longer(all_of(tmp_ids), values_to="staxid") %>%
  filter(!is.na(staxid)) %>%
  left_join(multi_taxids_taxonomy_df,by="staxid")

head(multi_taxids_df)
```
<br>

```{r}
multi2_lca <- custom.lca(multi_taxids_df, cutoff = 0.9)
```
<br>

split consensus taxonomy into multiple columns, to merge with the rest of the data
```{r}
multi2_lca_df <- multi2_lca %>%
  separate(consensus, c("kingdom", "phylum", "class", "order", "family", "genus", "species"), sep = '%') %>% 
  rename(representative = qseqid)

multi2_lca_df[multi2_lca_df == "NA" ] <- NA                                 # make sure "NA" is interpreted as NA
multi2_lca_df <- multi2_lca_df[rowSums(is.na(multi2_lca_df[ , 2:8])) != 7,]  # remove any rows with no taxonomic information (all NAs)
```
<br>

get the lowest level of taxonomy available for each hash using the function `coalesce`.
```{r}
multi2_lca_df %<>% 
  mutate(taxon = coalesce(species,genus,family,order,class,phylum,kingdom))
```
<br>

re-format the data frame so that the lowest taxonomic level available is reported in the column 'rank', and the 'score' and 'taxID' columns are saved as *consensus*. Add back in the dataset information
```{r}
multi2_lca_df %<>%
  group_by (representative) %>% nest %>% 
  mutate (rank = map_chr(data, ~pivot_longer(.x,cols = 1:7, names_to  = "rank", values_to ="name") %>% 
                           filter(name == taxon) %>% pull(rank))) %>% 
  select(representative, rank) %>% 
  inner_join(multi2_lca_df) %>%
  mutate(taxID="consensus", score="consensus")

multi2_lca_df %<>% left_join(dplyr::select(blast_to_process2, qseqid,dataset),by=c("representative"="qseqid"))

colnames(multi2_lca_df)
```
<br>



### conduct LCA for multiple sseqids

extract all of the taxa IDs 
```{r}
multi_seqids <- blast_to_process %>%
  distinct(staxid) %>%
  pull()
```
<br>

Use insect's database to get the associated taxonomy. Coerce the output (a list) into a dataframe.
```{r}
multi_seqids_taxonomy <- insect::get_lineage(as.numeric(multi_seqids), worlds.taxonomy)

multi_seqids_taxonomy_df <- tibble (staxid = multi_seqids,
                         taxonomy = multi_seqids_taxonomy) %>%
  mutate (Kingdom = map_chr(taxonomy, ~.x["kingdom"]),
          Phylum = map_chr(taxonomy, ~.x["phylum"]),
          Class = map_chr(taxonomy, ~.x["class"]),
          Order = map_chr(taxonomy, ~.x["order"]),
          Family= map_chr(taxonomy, ~.x["family"]),
          Genus = map_chr(taxonomy, ~.x["genus"]),
          Species = map_chr(taxonomy, ~.x["species"]))
```
<br>

Attach the taxonomy information to the hashes that were assigned multiple taxIDs. The resulting data frame has one row per unique taxID per hash.
```{r}
multi_seqids_df <- blast_to_process %>%
  left_join(multi_seqids_taxonomy_df,by="staxid")

colnames(multi_seqids_df)
```
<br>

```{r}
multi_lca <- custom.lca(multi_seqids_df, cutoff = 0.9)
```
<br>

split consensus taxonomy into multiple columns, to merge with the rest of the data
```{r}
multi_lca_df <- multi_lca %>%
  separate(consensus, c("kingdom", "phylum", "class", "order", "family", "genus", "species"), sep = '%') %>% 
  rename(representative = qseqid)

multi_lca_df[multi_lca_df == "NA" ] <- NA                                 # make sure "NA" is interpreted as NA
multi_lca_df <- multi_lca_df[rowSums(is.na(multi_lca_df[ , 2:8])) != 7,]  # remove any rows with no taxonomic information (all NAs)
```
<br>

get the lowest level of taxonomy available for each hash using the function `coalesce`.
```{r}
multi_lca_df %<>% 
  mutate(taxon = coalesce(species,genus,family,order,class,phylum,kingdom))
```
<br>

re-format the data frame so that the lowest taxonomic level available is reported in the column 'rank', and the 'score' and 'taxID' columns are saved as *consensus*. Add back in the dataset information
```{r}
multi_lca_df %<>%
  group_by (representative) %>% nest %>% 
  mutate (rank = map_chr(data, ~pivot_longer(.x,cols = 1:7, names_to  = "rank", values_to ="name") %>% 
                           filter(name == taxon) %>% pull(rank))) %>% 
  select(representative, rank) %>% 
  inner_join(multi_lca_df) %>%
  mutate(taxID="consensus", score="consensus")

multi_lca_df %<>% left_join(dplyr::select(blast_to_process, qseqid,dataset) %>% distinct(),by=c("representative"="qseqid"))

colnames(multi_lca_df)
```
<br>

combine with the data set from previous section, and write out both.
```{r}
multi_lca_df %<>% bind_rows(multi2_lca_df)

write.csv(multi_lca_df, here(outdir, consensus_out), row.names=FALSE)
```
<br>


### summarize taxonomy for unique staxids

Use the insect `get_lineage` function to attach taxonomy information to each unique taxID in the blast data.
```{r}
blast_taxonomy <- insect::get_lineage(unique(blast_dat$Ntaxid), worlds.taxonomy)

taxonomy_df <- tibble (staxid = unique(blast_dat$Ntaxid),
                         taxonomy = blast_taxonomy) %>%
  mutate (Kingdom = map_chr(taxonomy, ~.x["kingdom"]),
          Phylum = map_chr(taxonomy, ~.x["phylum"]),
          Class = map_chr(taxonomy, ~.x["class"]),
          Order = map_chr(taxonomy, ~.x["order"]),
          Family= map_chr(taxonomy, ~.x["family"]),
          Genus = map_chr(taxonomy, ~.x["genus"]),
          Species = map_chr(taxonomy, ~.x["species"]))


blast_taxonomy_df <- left_join(blast_dat, taxonomy_df,by=c("Ntaxid"="staxid"))
```
<br>


re-format the data frame so that the lowest taxonomic level available is reported in the columns 'taxon' and 'rank'. 
```{r}
blast_taxonomy_df %<>%
  clean_names() %>%
  mutate(taxon = coalesce(species,genus,family,order,class,phylum,kingdom)) 

blast_taxonomy_rank <- blast_taxonomy_df %>%
  pivot_longer(cols = c(species,genus,family,order,class,phylum,kingdom), names_to  = "rank", values_to ="name") %>% 
  filter(name == taxon) %>% pull(rank)

blast_taxonomy_df$rank <- blast_taxonomy_rank

colnames(blast_taxonomy_df)
```
<br>

save to file
```{r}
blast_taxonomy_df %<>% dplyr::select(-taxonomy)
write.csv(blast_taxonomy_df, here(outdir, unique_out), row.names=FALSE)
```
<br>
<br>




# Leray

Files containing Blast output.
```{r}
blast_filenames     <- c('mf.lane1.leray.merged.hashes.to.blast.txt',
                         'mf.lane1.leray.forward.hashes.to.blast.txt',
                         'mf.lane1.leray.reverse.hashes.to.blast.txt')
filename_identifier <- c('merged','forward','reverse')
```
<br>

Output files
```{r}
consensus_out <- 'lane1_leray_blast_multipleIDs_taxonomy.csv'
unique_out    <- 'lane1_leray_blast_uniqueIDs_taxonomy.csv'
all_out       <- 'lane1_leray_blast_taxonomy.csv'
```
<br>


## Read in data

The blast results should have the headers specified, following script 4. 

The column `staxid` contains the taxa IDs assigned to the DNA sequences (identified using their hashes); a hash can be assigned a single taxon ID (yay!) or multiple taxa IDs. When we coerce the column `staxid` into a numeric format, the hashes assigned to multiple taxa IDs will go to NA. We can then filter the data set to separate the hashes assigned to multiple taxa ids and explore. 

```{r data}
for(i in seq(1,length(blast_filenames))){
  tmp_file <- read_delim(here(indir,blast_filenames[i]),col_names = c("qseqid", "sseqid", "sacc", "pident", "length", "mismatch", "gapopen", "qcovus", "qstart", "qend", "sstart", "send", "evalue", "bitscore", "staxid", "qlen", "sscinames", "sseq"), delim = "\t") %>% mutate(Ntaxid=as.numeric(staxid))
  # split data by whether or not there are multiple IDs in the `staxid` column
  tmp_NA <- tmp_file %>%
    filter(is.na(Ntaxid)) %>%
    mutate(dataset = filename_identifier[i])
  tmp_dat <- tmp_file  %>%
    filter(!is.na(Ntaxid)) %>%
    mutate(dataset = filename_identifier[i])
  # split data again, by whether or not there are multiple rows / `sseqid` per `qseqid`
  tmp_multi <- tmp_dat %>%
    group_by(qseqid) %>%
    dplyr::filter(n() != 1) %>%
    ungroup()
  tmp_dat %<>% group_by(qseqid) %>%
    dplyr::filter(n() == 1) %>%
    ungroup()
  
  if(i==1){
    blast_dat <- tmp_dat
    blast_to_process  <- tmp_multi
    blast_to_process2 <- tmp_NA
  } else{
    blast_dat %<>% bind_rows(tmp_dat)
    blast_to_process  %<>% bind_rows(tmp_multi)
    blast_to_process2 %<>% bind_rows(tmp_NA)
  }
  
}
rm(tmp_file, tmp_dat, tmp_multi, tmp_NA)

unique(blast_dat$dataset)
dim(blast_dat)
dim(blast_to_process)
dim(blast_to_process2)
```
<br>
<br>

## Process results

### re-format multiple staxids & conduct LCA

extract all of the taxa IDs 
```{r}
multi_taxids <- blast_to_process2 %>%
  distinct(staxid) %>%
  pull()
multi_taxids <- unlist(sapply(multi_taxids,str_split,";")); names(multi_taxids) <- NULL
multi_taxids
```
<br>

Use insect's database to get the associated taxonomy. Coerce the output (a list) into a dataframe.
```{r}
multi_taxids_taxonomy <- insect::get_lineage(as.numeric(multi_taxids), worlds.taxonomy)

multi_taxids_taxonomy_df <- tibble (staxid = multi_taxids,
                         taxonomy = multi_taxids_taxonomy) %>%
  mutate (Kingdom = map_chr(taxonomy, ~.x["kingdom"]),
          Phylum = map_chr(taxonomy, ~.x["phylum"]),
          Class = map_chr(taxonomy, ~.x["class"]),
          Order = map_chr(taxonomy, ~.x["order"]),
          Family= map_chr(taxonomy, ~.x["family"]),
          Genus = map_chr(taxonomy, ~.x["genus"]),
          Species = map_chr(taxonomy, ~.x["species"]))
```
<br>

Attach the taxonomy information to the hashes that were assigned multiple taxIDs. The resulting data frame has one row per unique taxID per hash.
```{r}
tmp_ids <- c("id1","id2","id3","id4","id5","id6","id7","id8","id9","id10")
multi_taxids_df <- blast_to_process2 %>%
  separate(staxid, all_of(tmp_ids), fill = "right") %>%
  pivot_longer(all_of(tmp_ids), values_to="staxid") %>%
  filter(!is.na(staxid)) %>%
  left_join(multi_taxids_taxonomy_df,by="staxid")

head(multi_taxids_df)
```
<br>

```{r}
multi2_lca <- custom.lca(multi_taxids_df, cutoff = 0.9)
```
<br>

split consensus taxonomy into multiple columns, to merge with the rest of the data
```{r}
multi2_lca_df <- multi2_lca %>%
  separate(consensus, c("kingdom", "phylum", "class", "order", "family", "genus", "species"), sep = '%') %>% 
  rename(representative = qseqid)

multi2_lca_df[multi2_lca_df == "NA" ] <- NA                                 # make sure "NA" is interpreted as NA
multi2_lca_df <- multi2_lca_df[rowSums(is.na(multi2_lca_df[ , 2:8])) != 7,]  # remove any rows with no taxonomic information (all NAs)
```
<br>

get the lowest level of taxonomy available for each hash using the function `coalesce`.
```{r}
multi2_lca_df %<>% 
  mutate(taxon = coalesce(species,genus,family,order,class,phylum,kingdom))
```
<br>

re-format the data frame so that the lowest taxonomic level available is reported in the column 'rank', and the 'score' and 'taxID' columns are saved as *consensus*. Add back in the dataset information
```{r}
multi2_lca_df %<>%
  group_by (representative) %>% nest %>% 
  mutate (rank = map_chr(data, ~pivot_longer(.x,cols = 1:7, names_to  = "rank", values_to ="name") %>% 
                           filter(name == taxon) %>% pull(rank))) %>% 
  select(representative, rank) %>% 
  inner_join(multi2_lca_df) %>%
  mutate(taxID="consensus", score="consensus")

multi2_lca_df %<>% left_join(dplyr::select(blast_to_process2, qseqid,dataset),by=c("representative"="qseqid"))

colnames(multi2_lca_df)
```
<br>



### conduct LCA for multiple sseqids

extract all of the taxa IDs 
```{r}
multi_seqids <- blast_to_process %>%
  distinct(staxid) %>%
  pull()
```
<br>

Use insect's database to get the associated taxonomy. Coerce the output (a list) into a dataframe.
```{r}
multi_seqids_taxonomy <- insect::get_lineage(as.numeric(multi_seqids), worlds.taxonomy)

multi_seqids_taxonomy_df <- tibble (staxid = multi_seqids,
                         taxonomy = multi_seqids_taxonomy) %>%
  mutate (Kingdom = map_chr(taxonomy, ~.x["kingdom"]),
          Phylum = map_chr(taxonomy, ~.x["phylum"]),
          Class = map_chr(taxonomy, ~.x["class"]),
          Order = map_chr(taxonomy, ~.x["order"]),
          Family= map_chr(taxonomy, ~.x["family"]),
          Genus = map_chr(taxonomy, ~.x["genus"]),
          Species = map_chr(taxonomy, ~.x["species"]))
```
<br>

Attach the taxonomy information to the hashes that were assigned multiple taxIDs. The resulting data frame has one row per unique taxID per hash.
```{r}
multi_seqids_df <- blast_to_process %>%
  left_join(multi_seqids_taxonomy_df,by="staxid")

colnames(multi_seqids_df)
```
<br>

```{r}
multi_lca <- custom.lca(multi_seqids_df, cutoff = 0.9)
```
<br>

split consensus taxonomy into multiple columns, to merge with the rest of the data
```{r}
multi_lca_df <- multi_lca %>%
  separate(consensus, c("kingdom", "phylum", "class", "order", "family", "genus", "species"), sep = '%') %>% 
  rename(representative = qseqid)

multi_lca_df[multi_lca_df == "NA" ] <- NA                                 # make sure "NA" is interpreted as NA
multi_lca_df <- multi_lca_df[rowSums(is.na(multi_lca_df[ , 2:8])) != 7,]  # remove any rows with no taxonomic information (all NAs)
```
<br>

get the lowest level of taxonomy available for each hash using the function `coalesce`.
```{r}
multi_lca_df %<>% 
  mutate(taxon = coalesce(species,genus,family,order,class,phylum,kingdom))
```
<br>

re-format the data frame so that the lowest taxonomic level available is reported in the column 'rank', and the 'score' and 'taxID' columns are saved as *consensus*. Add back in the dataset information
```{r}
multi_lca_df %<>%
  group_by (representative) %>% nest %>% 
  mutate (rank = map_chr(data, ~pivot_longer(.x,cols = 1:7, names_to  = "rank", values_to ="name") %>% 
                           filter(name == taxon) %>% pull(rank))) %>% 
  select(representative, rank) %>% 
  inner_join(multi_lca_df) %>%
  mutate(taxID="consensus", score="consensus")

multi_lca_df %<>% left_join(dplyr::select(blast_to_process, qseqid,dataset) %>% distinct(),by=c("representative"="qseqid"))

colnames(multi_lca_df)
```
<br>

combine with the data set from previous section, and write out both.
```{r}
multi_lca_df %<>% bind_rows(multi2_lca_df)

write.csv(multi_lca_df, here(outdir, consensus_out), row.names=FALSE)
```
<br>


### summarize taxonomy for unique staxids

Use the insect `get_lineage` function to attach taxonomy information to each unique taxID in the blast data.
```{r}
blast_taxonomy <- insect::get_lineage(unique(blast_dat$Ntaxid), worlds.taxonomy)

taxonomy_df <- tibble (staxid = unique(blast_dat$Ntaxid),
                         taxonomy = blast_taxonomy) %>%
  mutate (Kingdom = map_chr(taxonomy, ~.x["kingdom"]),
          Phylum = map_chr(taxonomy, ~.x["phylum"]),
          Class = map_chr(taxonomy, ~.x["class"]),
          Order = map_chr(taxonomy, ~.x["order"]),
          Family= map_chr(taxonomy, ~.x["family"]),
          Genus = map_chr(taxonomy, ~.x["genus"]),
          Species = map_chr(taxonomy, ~.x["species"]))


blast_taxonomy_df <- left_join(blast_dat, taxonomy_df,by=c("Ntaxid"="staxid"))
```
<br>


re-format the data frame so that the lowest taxonomic level available is reported in the columns 'taxon' and 'rank'. 
```{r}
blast_taxonomy_df %<>%
  clean_names() %>%
  mutate(taxon = coalesce(species,genus,family,order,class,phylum,kingdom)) 

blast_taxonomy_rank <- blast_taxonomy_df %>%
  pivot_longer(cols = c(species,genus,family,order,class,phylum,kingdom), names_to  = "rank", values_to ="name") %>% 
  filter(name == taxon) %>% pull(rank)

blast_taxonomy_df$rank <- blast_taxonomy_rank

colnames(blast_taxonomy_df)
```
<br>

save to file
```{r}
blast_taxonomy_df %<>% dplyr::select(-taxonomy)
write.csv(blast_taxonomy_df, here(outdir, unique_out), row.names=FALSE)
```
<br>
<br>





# LerayXT

Files containing Blast output.
```{r}
blast_filenames     <- c('mf.lane1.lerayxt.merged.hashes.to.blast.txt',
                         'mf.lane1.lerayxt.forward.hashes.to.blast.txt',
                         'mf.lane1.lerayxt.reverse.hashes.to.blast.txt')
filename_identifier <- c('merged','forward','reverse')
```
<br>

Output files
```{r}
consensus_out <- 'lane1_lerayxt_blast_multipleIDs_taxonomy.csv'
unique_out    <- 'lane1_lerayxt_blast_uniqueIDs_taxonomy.csv'
all_out       <- 'lane1_lerayxt_blast_taxonomy.csv'
```
<br>


## Read in data

The blast results should have the headers specified, following script 4. 

The column `staxid` contains the taxa IDs assigned to the DNA sequences (identified using their hashes); a hash can be assigned a single taxon ID (yay!) or multiple taxa IDs. When we coerce the column `staxid` into a numeric format, the hashes assigned to multiple taxa IDs will go to NA. We can then filter the data set to separate the hashes assigned to multiple taxa ids and explore. 

```{r data}
for(i in seq(1,length(blast_filenames))){
  tmp_file <- read_delim(here(indir,blast_filenames[i]),col_names = c("qseqid", "sseqid", "sacc", "pident", "length", "mismatch", "gapopen", "qcovus", "qstart", "qend", "sstart", "send", "evalue", "bitscore", "staxid", "qlen", "sscinames", "sseq"), delim = "\t", col_types="cccdddddddddddcccc") %>% mutate(Ntaxid=as.numeric(staxid))
  # split data by whether or not there are multiple IDs in the `staxid` column
  tmp_NA <- tmp_file %>%
    filter(is.na(Ntaxid)) %>%
    mutate(dataset = filename_identifier[i])
  tmp_dat <- tmp_file  %>%
    filter(!is.na(Ntaxid)) %>%
    mutate(dataset = filename_identifier[i])
  # split data again, by whether or not there are multiple rows / `sseqid` per `qseqid`
  tmp_multi <- tmp_dat %>%
    group_by(qseqid) %>%
    dplyr::filter(n() != 1) %>%
    ungroup()
  tmp_dat %<>% group_by(qseqid) %>%
    dplyr::filter(n() == 1) %>%
    ungroup()
  
  if(i==1){
    blast_dat <- tmp_dat
    blast_to_process  <- tmp_multi
    blast_to_process2 <- tmp_NA
  } else{
    # blast_dat$staxid <- as.character(blast_dat$staxid)
    blast_dat %<>% bind_rows(tmp_dat)
    # blast_to_process$staxid <- as.character(blast_to_process$staxid)
    blast_to_process  %<>% bind_rows(tmp_multi)
    # blast_to_process2$staxid <- as.character(blast_to_process2$staxid)
    blast_to_process2 %<>% bind_rows(tmp_NA)
  }
  
}
rm(tmp_file, tmp_dat, tmp_multi, tmp_NA)

unique(blast_dat$dataset)
dim(blast_dat)
dim(blast_to_process)
dim(blast_to_process2)
```
<br>
<br>

## Process results

### re-format multiple staxids & conduct LCA

extract all of the taxa IDs 
```{r}
multi_taxids <- blast_to_process2 %>%
  distinct(staxid) %>%
  pull()
multi_taxids <- unlist(sapply(multi_taxids,str_split,";")); names(multi_taxids) <- NULL
multi_taxids
```
<br>

Use insect's database to get the associated taxonomy. Coerce the output (a list) into a dataframe.
```{r}
multi_taxids_taxonomy <- insect::get_lineage(as.numeric(multi_taxids), worlds.taxonomy)

multi_taxids_taxonomy_df <- tibble (staxid = multi_taxids,
                         taxonomy = multi_taxids_taxonomy) %>%
  mutate (Kingdom = map_chr(taxonomy, ~.x["kingdom"]),
          Phylum = map_chr(taxonomy, ~.x["phylum"]),
          Class = map_chr(taxonomy, ~.x["class"]),
          Order = map_chr(taxonomy, ~.x["order"]),
          Family= map_chr(taxonomy, ~.x["family"]),
          Genus = map_chr(taxonomy, ~.x["genus"]),
          Species = map_chr(taxonomy, ~.x["species"]))
```
<br>

Attach the taxonomy information to the hashes that were assigned multiple taxIDs. The resulting data frame has one row per unique taxID per hash.
```{r}
tmp_ids <- c("id1","id2","id3","id4","id5","id6","id7","id8","id9","id10")
multi_taxids_df <- blast_to_process2 %>%
  separate(staxid, all_of(tmp_ids), fill = "right") %>%
  pivot_longer(all_of(tmp_ids), values_to="staxid") %>%
  filter(!is.na(staxid)) %>%
  left_join(multi_taxids_taxonomy_df,by="staxid")

head(multi_taxids_df)
```
<br>

```{r}
multi2_lca <- custom.lca(multi_taxids_df, cutoff = 0.9)
```
<br>

split consensus taxonomy into multiple columns, to merge with the rest of the data
```{r}
multi2_lca_df <- multi2_lca %>%
  separate(consensus, c("kingdom", "phylum", "class", "order", "family", "genus", "species"), sep = '%') %>% 
  rename(representative = qseqid)

multi2_lca_df[multi2_lca_df == "NA" ] <- NA                                 # make sure "NA" is interpreted as NA
multi2_lca_df <- multi2_lca_df[rowSums(is.na(multi2_lca_df[ , 2:8])) != 7,]  # remove any rows with no taxonomic information (all NAs)
```
<br>

get the lowest level of taxonomy available for each hash using the function `coalesce`.
```{r}
multi2_lca_df %<>% 
  mutate(taxon = coalesce(species,genus,family,order,class,phylum,kingdom))
```
<br>

re-format the data frame so that the lowest taxonomic level available is reported in the column 'rank', and the 'score' and 'taxID' columns are saved as *consensus*. Add back in the dataset information
```{r}
multi2_lca_df %<>%
  group_by (representative) %>% nest %>% 
  mutate (rank = map_chr(data, ~pivot_longer(.x,cols = 1:7, names_to  = "rank", values_to ="name") %>% 
                           filter(name == taxon) %>% pull(rank))) %>% 
  select(representative, rank) %>% 
  inner_join(multi2_lca_df) %>%
  mutate(taxID="consensus", score="consensus")

multi2_lca_df %<>% left_join(dplyr::select(blast_to_process2, qseqid,dataset),by=c("representative"="qseqid"))

colnames(multi2_lca_df)
```
<br>



### conduct LCA for multiple sseqids

extract all of the taxa IDs 
```{r}
multi_seqids <- blast_to_process %>%
  distinct(staxid) %>%
  pull()
```
<br>

Use insect's database to get the associated taxonomy. Coerce the output (a list) into a dataframe.
```{r}
multi_seqids_taxonomy <- insect::get_lineage(as.numeric(multi_seqids), worlds.taxonomy)

multi_seqids_taxonomy_df <- tibble (staxid = multi_seqids,
                         taxonomy = multi_seqids_taxonomy) %>%
  mutate (Kingdom = map_chr(taxonomy, ~.x["kingdom"]),
          Phylum = map_chr(taxonomy, ~.x["phylum"]),
          Class = map_chr(taxonomy, ~.x["class"]),
          Order = map_chr(taxonomy, ~.x["order"]),
          Family= map_chr(taxonomy, ~.x["family"]),
          Genus = map_chr(taxonomy, ~.x["genus"]),
          Species = map_chr(taxonomy, ~.x["species"]))
```
<br>

Attach the taxonomy information to the hashes that were assigned multiple taxIDs. The resulting data frame has one row per unique taxID per hash.
```{r}
multi_seqids_df <- blast_to_process %>%
  left_join(multi_seqids_taxonomy_df,by="staxid")

colnames(multi_seqids_df)
```
<br>

```{r}
multi_lca <- custom.lca(multi_seqids_df, cutoff = 0.9)
```
<br>

split consensus taxonomy into multiple columns, to merge with the rest of the data
```{r}
multi_lca_df <- multi_lca %>%
  separate(consensus, c("kingdom", "phylum", "class", "order", "family", "genus", "species"), sep = '%') %>% 
  rename(representative = qseqid)

multi_lca_df[multi_lca_df == "NA" ] <- NA                                 # make sure "NA" is interpreted as NA
multi_lca_df <- multi_lca_df[rowSums(is.na(multi_lca_df[ , 2:8])) != 7,]  # remove any rows with no taxonomic information (all NAs)
```
<br>

get the lowest level of taxonomy available for each hash using the function `coalesce`.
```{r}
multi_lca_df %<>% 
  mutate(taxon = coalesce(species,genus,family,order,class,phylum,kingdom))
```
<br>

re-format the data frame so that the lowest taxonomic level available is reported in the column 'rank', and the 'score' and 'taxID' columns are saved as *consensus*. Add back in the dataset information
```{r}
multi_lca_df %<>%
  group_by (representative) %>% nest %>% 
  mutate (rank = map_chr(data, ~pivot_longer(.x,cols = 1:7, names_to  = "rank", values_to ="name") %>% 
                           filter(name == taxon) %>% pull(rank))) %>% 
  select(representative, rank) %>% 
  inner_join(multi_lca_df) %>%
  mutate(taxID="consensus", score="consensus")

multi_lca_df %<>% left_join(dplyr::select(blast_to_process, qseqid,dataset) %>% distinct(),by=c("representative"="qseqid"))

colnames(multi_lca_df)
```
<br>

combine with the data set from previous section, and write out both.
```{r}
multi_lca_df %<>% bind_rows(multi2_lca_df)

write.csv(multi_lca_df, here(outdir, consensus_out), row.names=FALSE)
```
<br>


### summarize taxonomy for unique staxids

Use the insect `get_lineage` function to attach taxonomy information to each unique taxID in the blast data.
```{r}
blast_taxonomy <- insect::get_lineage(unique(blast_dat$Ntaxid), worlds.taxonomy)

taxonomy_df <- tibble (staxid = unique(blast_dat$Ntaxid),
                         taxonomy = blast_taxonomy) %>%
  mutate (Kingdom = map_chr(taxonomy, ~.x["kingdom"]),
          Phylum = map_chr(taxonomy, ~.x["phylum"]),
          Class = map_chr(taxonomy, ~.x["class"]),
          Order = map_chr(taxonomy, ~.x["order"]),
          Family= map_chr(taxonomy, ~.x["family"]),
          Genus = map_chr(taxonomy, ~.x["genus"]),
          Species = map_chr(taxonomy, ~.x["species"]))


blast_taxonomy_df <- left_join(blast_dat, taxonomy_df,by=c("Ntaxid"="staxid"))
```
<br>


re-format the data frame so that the lowest taxonomic level available is reported in the columns 'taxon' and 'rank'. 
```{r}
blast_taxonomy_df %<>%
  clean_names() %>%
  mutate(taxon = coalesce(species,genus,family,order,class,phylum,kingdom)) 

blast_taxonomy_rank <- blast_taxonomy_df %>%
  pivot_longer(cols = c(species,genus,family,order,class,phylum,kingdom), names_to  = "rank", values_to ="name") %>% 
  filter(name == taxon) %>% pull(rank)

blast_taxonomy_df$rank <- blast_taxonomy_rank

colnames(blast_taxonomy_df)
```
<br>

save to file
```{r}
blast_taxonomy_df %<>% dplyr::select(-taxonomy)
write.csv(blast_taxonomy_df, here(outdir, unique_out), row.names=FALSE)
```
<br>
<br>
